remove unneccessary controller
