+++ src/app.js
@@ -67,7 +67,7 @@ var usersRoutes = require('../controllers/usersController')
 var adminUsersRoutes = require('../controllers/adminUsersController')
 var generalUsersRoutes = require('../controllers/generalUsersController')
 var pollsRoutes = require('../controllers/pollsController')
-var managePollsRoutes = require('../controllers/managePollsController')
+//var managePollsRoutes = require('../controllers/managePollsController')
 var myPollsRoutes = require('../controllers/myPollsController')
 
 
@@ -77,7 +77,7 @@ app.use('/users', usersRoutes)
 app.use('/adminUsers', adminUsersRoutes)
 app.use('/generalUsers', generalUsersRoutes)
 app.use('/polls', pollsRoutes)
-app.use('/polls', managePollsRoutes)
+//app.use('/polls', managePollsRoutes)
 app.use('/mypolls', myPollsRoutes)
