+++ views/layouts/main.handlebars
@@ -22,7 +22,7 @@
             <a class="nav-link" href="/polls">Create Polls</a> <!-- admin route -->
           </li>
           <li class="nav-item">
-            <a class="nav-link" href="/polls/manage">Manage Polls</a> <!-- admin route -->
+            <a class="nav-link" href="/polls/list">Manage Polls</a> <!-- admin route -->
           </li>
         {{else}}
           <li class="nav-item">
